test: extract tests for setting AllowedERC725YKeys in separate file
